+++ src/pydiverse/transform/_internal/ops/registry.py
